gobject: Fix -Wshorten-64-to-32 warnings and enable warning flag

See merge request GNOME/glib!4950

+++ gobject/gbinding.c
@@ -256,8 +256,8 @@ struct _GBinding
 
   GBindingFlags flags;
 
-  guint source_notify; /* LOCK: unbind_lock */
-  guint target_notify; /* LOCK: unbind_lock */
+  unsigned long source_notify; /* LOCK: unbind_lock */
+  unsigned long target_notify; /* LOCK: unbind_lock */
   gboolean target_weak_notify_installed; /* LOCK: unbind_lock */
 
   /* a guard, to avoid loops */
